backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
yt-dlp==2026.3.17
slowapi==0.1.9
python-dotenv==1.0.1
httpx[http2]==0.28.1
